requirements.txt
asgiref==3.7.2
Django==4.2.6
django-address==0.2.8
django-phonenumber-field==7.2.0
phonenumbers==8.13.22
Pillow==10.0.1
sqlparse==0.4.4
tzdata==2023.3
